test: combine PeopleService test files into one file
